Initial transition to use skopeo_worker on the getsizeworker
